# specs/002-users-seohun-documents/contracts/api-spec.yaml
openapi: 3.0.0
info:
  title: AgentTrust API
  description: XRPL-based AI Agent License Marketplace API
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Development server
paths:
...
  /agents:
    get:
      summary: List all AI agents
      tags:
        - Marketplace
      parameters:
        - in: query
          name: category
          schema:
            type: string
            enum: [popular, new, nlp, computer_vision, reinforcement]
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: List of AI agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/AIAgent'
                  total:
                    type: integer

    post:
      summary: Register new AI agent
      tags:
        - Developer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - description
                - category
                - priceXRP
                - ipfsHash
              properties:
                name:
                  type: string
                description:
                  type: string
                category:
                  type: string
                  enum: [NLP, Computer Vision, RL, Other]
                priceXRP:
                  type: number
                  minimum: 0.1
                ipfsHash:
                  type: string
                  pattern: '^Qm[a-zA-Z0-9]{44}$'
                imageUrl:
                  type: string
      responses:
        201:
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAgent'
  /agents/{agentId}:
    get:
      summary: Get AI agent details
      tags:
        - Marketplace
      parameters:
        - in: path
          name: agentId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIAgentDetail'
...
  /downloads/{token}:
    get:
      summary: Download AI agent with token
      tags:
        - Downloads
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        200:
          description: File download
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        403:
          description: Invalid or expired token
  /downloads/request:
    post:
      summary: Request new download token
      tags:
        - Downloads
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - licenseId
              properties:
                licenseId:
                  type: string
      responses:
        200:
          description: Download token generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expiresAt:
                    type: string
                    format: date-time
                  downloadUrl:
                    type: string
...
components:
  schemas:
...
    AIAgent:
      type: object
      properties:
        agentId:
          type: string
        nftId:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        priceXRP:
          type: number
        imageUrl:
          type: string
        totalSales:
          type: integer
        averageRating:
          type: number
        createdAt:
          type: string
          format: date-time
    AIAgentDetail:
      allOf:
        - $ref: '#/components/schemas/AIAgent'
        - type: object
          properties:
            did:
              type: string
            credentialType:
              type: string
            ownerWallet:
              type: string
            reviews:
              type: array
              items:
                $ref: '#/components/schemas/Review'
...
    Review:
      type: object
      properties:
        reviewId:
          type: string
        agentId:
          type: string
        licenseId:
          type: string
        reviewerWallet:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT